modify:
- header-only version of matrix and vector complete
