--- simutil/simutil_includes.h
@@ -1,7 +1,0 @@
-#ifndef SIMUTILS_INCLUDES_H
-#define SIMUTILS_INCLUDES_H
-
-#include <stdio.h>
-#include <stdlib.h>
-
-#endif
