--- simutil/simutils.h
@@ -1,7 +1,0 @@
-#ifndef SIMUTILS_H
-#define SIMUTILS_H
-
-#include "matrix.h"
-#include "vector.h"
-
-#endif
